requirements.txt
fastapi
uvicorn
qrcode
pillow
pydantic-settings
sqlalchemy>=2
pymysql
